tests/mediacapture-depth/MediaStreamTrack-end-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test that mediastreamtrack are properly ended</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#life-cycle">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Description</h2>
<p>
  This test checks that the video tracks of
MediaStream object returned by the success callback in getUserMedia are
correctly set into inactive state when permission is revoked.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external).
  </li>
  <li>
    When prompted, accept to share your depth stream, and then revoke that permission.
  </li>
</ol>

<div id="log"></div>

<script>

async_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    let vidTrack = stream.getVideoTracks()[0];
    vidTrack.onended = t.step_timeout(() => {
      assert_equals(vidTrack.readyState, "ended", "Video track has been ended as expected");
      assert_false(stream.active, "MediaStream has been inactive as expected");
      t.done();
    }, 5000);
    assert_equals(vidTrack.readyState, "live", "The video track object is in live state");
  });// longer timeout since requires user interaction
}, "Tests that the video MediaStreamTrack objects are properly ended on permission revocation", {timeout: 20000});

</script>
